bfpInventoryBackend/APIs/requirements.txt
Flask
flask-cors
bcrypt
PyJWT
psycopg2-binary
